+++ mitra_api/src/web_client/mod.rs
@@ -1,3 +1,4 @@
 mod types;
 pub mod urls;
+mod utils;
 pub mod views;