+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Set `discoverable` flag to `true` on Mastodon API `Account` entities.
+- Replace `description` meta tag on post and profile pages.
 
 ### Changed